requirements/apps.txt
Django==1.3
amqplib==1.0.0
anyjson==0.3.1
celery==2.3.1
distribute==0.6.15
django-celery==2.3.3
django-extensions==0.6
django-picklefield==0.1.9
kombu==1.2.1
paramiko==1.7.7.1
pycrypto==2.3
pyparsing==1.5.6
python-dateutil==1.5
wsgiref==0.1.2
